testing

+++ postage/test.sh
@@ -68,7 +68,7 @@ if test $(uname -s) = "Linux"; then
 		cp -f ./dist/*.rpm /mnt/Groups/wfprod_group/postage/packages/
 	}
 
-#elif test $(uname -s) = "Darwin"; then
+elif test $(uname -s) = "Darwin"; then
 	./dist/mac/Postage.app/Contents/MacOS/Postage --postage-test & export POSTAGEPID="$!"
 	printf "HTTP/1.1 200 OK\r\n\r\n\r\n" | ncat -l -p 45654
 	kill $POSTAGEPID
